// src/types/ThemeColors.ts
export interface ThemeColors {
  primary: string;
  background: string;
  text: {
    primary: string;
    secondary: string;
  };
  surface: {
    primary: string;
    secondary: string;
  };
  action: {
    active: string;
    disabled: string;
  };
  status: {
    error: string;
    success: string;
    warning: string;
  };
  utility: {
    transparent: string;
  };
}